Consolidate ping
