--- inodaqv2/command_ping.h
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace Command
-{
-void command_ping();
-}
