+++ inodaqv2/commands.hpp
@@ -5,3 +5,4 @@
 void command_aread();
 void command_dig(const ::String &command);
 void command_dread();
+void command_ping();
